+++ packages/dash-core-components/dash_core_components_base/__init__.py
@@ -80,10 +80,10 @@ _js_dist = [
         'dynamic': True
     },
     {
-        'relative_package_path': 'plotly-1.50.1.min.js',
+        'relative_package_path': 'plotly-1.51.1.min.js',
         'external_url': (
             'https://unpkg.com/dash-core-components@{}'
-            '/dash_core_components/plotly-1.50.1.min.js'
+            '/dash_core_components/plotly-1.51.1.min.js'
         ).format(__version__),
         'namespace': 'dash_core_components',
         'async': 'eager'